chore: split files a bit more
